+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+- Support Hummus protocol.
+
 ## [2.6.0] - 2023-05-09
 
 ### Added
 
-- Support Hummus protocol.
 - Support CompoundV3.
 
 ## [2.5.0] - 2023-03-16