+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [2.6.1] - 2023-05-24
+### Changed
+
+- Hotfix CompoundV3 handler.
+
 ## [2.6.0] - 2023-05-09
 
 ### Added